third_party/blink/web_tests/external/wpt/html/semantics/popovers/light-dismiss-event-ordering-expected.txt

This is a testharness.js-based test.
[FAIL] Tests the order of pointer/mouse events during popover light dismiss.
  assert_array_equals: pointer and popover events should be fired in the correct order. expected property 2 to be "beforetoggle newState: closed" but got "pointerup" (expected array ["pointerdown", "mousedown", "beforetoggle newState: closed", "pointerup", "mouseup", "click"] got ["pointerdown", "mousedown", "pointerup", "mouseup", "beforetoggle newState: closed", "click"])
Harness: the test ran to completion.
